+++ CHANGELOG.MD
@@ -4,6 +4,14 @@
 
 ---
 
+## [1.1.0] - 2025-03-30
+### Telegram Bot API Version 8.3
+### Added:
+- documentation for every package
+### Changes:
+- quick refactor: moving api/objects and api/methods to the root directory
+- swapped token and destination in SendRequest and SendRequestWith
+
 ## [1.0.0] - 2025-03-30
 ### Telegram Bot API Version 8.3
 ### First release
