+++ methods/doc.go
@@ -1,0 +1,5 @@
+// This package provides a set of methods as defined in Telegram Bot API documentation.
+// Each method implements [gotely.Method] and can be used with [gotely.SendRequest].
+//
+// Licensed under the MIT License. See LICENSE file for details.
+package methods
