+++ methods/inline.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/bigelle/gotely"
-	"github.com/bigelle/gotely/api/objects"
+	"github.com/bigelle/gotely/objects"
 )
 
 // Use this method to send answers to an inline query.
