+++ methods/passport.go
@@ -5,7 +5,7 @@ import (
 	"io"
 
 	"github.com/bigelle/gotely"
-	"github.com/bigelle/gotely/api/objects"
+	"github.com/bigelle/gotely/objects"
 )
 
 // /Informs a user that some of the Telegram Passport elements they provided contains errors.
