+++ objects/doc.go
@@ -1,0 +1,5 @@
+// This package provides a set of types defined in Telegram Bot API documentation.
+// Each type is ready to be used in API requests.
+//
+// Licensed under the MIT License. See LICENSE file for details.
+package objects
