+++ objects/inline.go
