+++ objects/stickers.go
