+++ tgbot/longpolling/doc.go
@@ -1,0 +1,6 @@
+// This package provides a way to set up a Long Polling bot.
+// This bot is using long polling to request for new updates from Telegram Bot API.
+// Every received update would be handled using method defined in [tgbot.Bot] implementation.
+//
+// Licensed under the MIT License. See LICENSE file for details.
+package longpolling
