+++ tgbot/webhook/doc.go
@@ -1,0 +1,7 @@
+// This package provides a way to set up a server
+// that will handle updates coming from Telegram Bot API,
+// as well as a set of types used to set or delete webhook.
+// Each type implements [gotely.Method] and can be used with [gotely.SendRequest]
+//
+// Licensed under the MIT License. See LICENSE file for details.
+package webhook
